Add remove_src option to enable copy

+++ cloudpathlib/gs/gsclient.py
@@ -158,7 +158,7 @@ class GSClient(Client):
 
             yield self.CloudPath(f"gs://{cloud_path.bucket}/{o.name}")
 
-    def _move_file(self, src: GSPath, dst: GSPath) -> GSPath:
+    def _move_file(self, src: GSPath, dst: GSPath, remove_src: bool = True) -> GSPath:
         # just a touch, so "REPLACE" metadata
         if src == dst:
             bucket = self.client.bucket(src.bucket)
@@ -177,7 +177,9 @@ class GSClient(Client):
 
             src_blob = src_bucket.get_blob(src.blob)
             src_bucket.copy_blob(src_blob, dst_bucket, dst.blob)
-            src_blob.delete()
+
+            if remove_src:
+                src_blob.delete()
 
         return dst
